+++ scripts/ci-env.ps1
@@ -11,7 +11,7 @@ switch ($name) {
         return "2022.3.62f2"
     }
     "unity6000.0" {
-        return "6000.0.58f1"
+        return "6000.0.60f1"
     }
     "unity6000.1" {
         return "6000.1.17f1"