+++ test/Scripts.Integration.Test/add-sentry.ps1
@@ -14,6 +14,6 @@ $UnityPath = FormatUnityPath $UnityPath
 RunUnityAndExpect $UnityPath "AddSentryPackage" "Sentry Package Installation: SUCCESS" @( `
         "-batchmode", "-projectPath ", "$(GetNewProjectPath)", "-installSentry", "Disk")
 
-Write-Host "Copying Integration Test Files"
+Write-Log "Copying Integration Test Files"
 New-Item -Path "$(GetNewProjectAssetsPath)" -Name "Scripts" -ItemType "directory"
 Copy-Item -Recurse "$IntegrationScriptsPath/Scripts/*" -Destination "$(GetNewProjectAssetsPath)/Scripts/"